Added aggregate pattern in DatabaseConnector
